:recycle: refactor : 챌린지 참여 기능 이동
Relate to: #11
