+++ src/main/java/com/example/groutine/domain/challenge/service/ChallengeMemberQueryService.java
@@ -1,0 +1,11 @@
+package com.example.groutine.domain.challenge.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+public class ChallengeMemberQueryService {
+}
